Refactor the rest of the app
