+++ CoreLocationComponent/CoreLocationComponent/Utilities/Constant/PackageConstant.swift
@@ -1,0 +1,10 @@
+//
+//  PackageConstant.swift
+//  CoreLocationComponent
+//
+//  Created by Nathanael Juan Gauthama on 16/07/24.
+//
+
+import Foundation
+
+let packageIdentifier: String = "group.com.pang.CoreLocationComponent"
