+++ CoreLocationComponent/CoreLocationComponent/ViewModel/Manager/WeatherManager.swift
@@ -13,12 +13,12 @@ import SwiftUI
 class WeatherManager: ObservableObject {
     
     static var shared = WeatherManager()
-    @AppStorage("safeWeatherData", store: UserDefaults(suiteName: "group.com.pang.CoreLocationComponent")) var safeWeatherData = " "
+    @AppStorage("safeWeatherData", store: UserDefaults(suiteName: packageIdentifier)) var safeWeatherData = " "
     
     @Published var weather: Weather?
     
     func getWeather(latitude: Double, longitude: Double) {
-        async {
+        Task.init{
             do {
                 weather = try await Task.detached(priority: .userInitiated) {
                     return try await WeatherService.shared.weather(for: .init(latitude: latitude, longitude: longitude))
@@ -114,49 +114,6 @@ class WeatherManager: ObservableObject {
         }
         
         
-    }
-    
-    struct GroupedWeather: Identifiable {
-        var id = UUID()
-        var type: WeatherType
-        var items: [GraphModel]
-    }
-
-    enum WeatherType {
-        case safe
-        case risky
-    }
-
-    
-    func groupWeatherData(_ weatherData: [GraphModel]) -> [GroupedWeather] {
-        var groupedWeather: [GroupedWeather] = []
-        
-        var currentType: WeatherType?
-        var currentItems: [GraphModel] = []
-        
-        for weatherItem in weatherData {
-            let isSafe = safeWeather.contains { $0.startTime <= weatherItem.value.date && $0.endTime >= weatherItem.value.date }
-            let weatherType: WeatherType = isSafe ? .safe : .risky
-            
-            if currentType == nil {
-                currentType = weatherType
-            }
-            
-            if currentType == weatherType {
-                currentItems.append(weatherItem)
-            } else {
-                groupedWeather.append(GroupedWeather(type: currentType!, items: currentItems))
-                currentType = weatherType
-                currentItems = [weatherItem]
-            }
-        }
-        
-        if let currentType = currentType {
-            groupedWeather.append(GroupedWeather(type: currentType, items: currentItems))
-        }
-        return groupedWeather
-    }
-
-    
+    }    
 }
